Add JWT util.

+++ socialWeb_user/src/main/java/com/socialWeb/user/service/UserService.java
@@ -1,4 +1,5 @@
 package com.socialWeb.user.service;
+import io.jsonwebtoken.Claims;
 import java.util.Date;
 import	java.util.HashMap;
 
@@ -15,6 +16,7 @@ import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
 
+import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,6 +32,7 @@ import util.IdWorker;
 
 import com.socialWeb.user.dao.UserDao;
 import com.socialWeb.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -54,6 +57,9 @@ public class UserService {
 
 	@Autowired
 	private BCryptPasswordEncoder encoder;
+
+	@Autowired
+	private HttpServletRequest request;
 
 	/**
 	 * 查询全部列表
@@ -126,6 +132,10 @@ public class UserService {
 	 * @param id
 	 */
 	public void deleteById(String id) {
+		String token = (String) request.getAttribute("claims_admin");
+		if (token == null || "".equals(token)){
+			throw new RuntimeException("权限不足");
+		}
 		userDao.deleteById(id);
 	}
